requirements.txt
nltk==3.3
pdfminer.six==20170720
